+++ R/coord_geo.R
@@ -1,7 +1,7 @@
 ## declare variables that are used within aes() to prevent
 ## R CMD check from complaining
 utils::globalVariables(c("min_age", "max_age", "mid_age", "label",
-                         "name", "translate"))
+                         "name", "translate", "stages"))
 
 #' Transformed coordinate system with geological timescale
 #'