+++ NEWS.md
@@ -6,6 +6,7 @@ deeptime now requires ggplot2 version 3.4.0 or higher. This version introduces n
 * Fixed `size = "auto"` when `center_end_labels = TRUE` in coord_geo()
 * Added `family` and `fontface` arguments to coord_geo()
 * Added the ability to include interval labels with coord_geo_polar() (#48)
+* Added facet_grid_color() and facet_wrap_color() for changing strip background colors (#50)
 
 # deeptime 1.0.1